Fix OSGi test
