+++ h2spatial/src/main/java/org/h2gis/h2spatial/internal/type/GeometryTypeNameFromConstraint.java
@@ -2,6 +2,7 @@ package org.h2gis.h2spatial.internal.type;
 
 import org.h2gis.h2spatialapi.DeterministicScalarFunction;
 import org.h2gis.utilities.GeometryTypeCodes;
+import org.h2gis.utilities.SFSUtilities;
 
 import java.lang.reflect.Field;
 import java.util.HashMap;
@@ -12,18 +13,6 @@ import java.util.Map;
  * @author Nicolas Fortin
  */
 public class GeometryTypeNameFromConstraint extends DeterministicScalarFunction {
-
-    private static final Map<Integer, String> TYPE_MAP = new HashMap<Integer, String>();
-    static {
-        // Cache GeometryTypeCodes into a static HashMap
-        for(Field field : GeometryTypeCodes.class.getDeclaredFields()) {
-            try {
-                TYPE_MAP.put(field.getInt(null),field.getName());
-            } catch (IllegalAccessException ex) {
-                //pass
-            }
-        }
-    }
 
     public GeometryTypeNameFromConstraint() {
         addProperty(PROP_REMARKS, "Parse the constraint and return the Geometry type name");
@@ -42,10 +31,6 @@ public class GeometryTypeNameFromConstraint extends DeterministicScalarFunction 
      */
     public static String getGeometryTypeNameFromConstraint(String constraint, int numericPrecision) {
         int geometryTypeCode = GeometryTypeFromConstraint.geometryTypeFromConstraint(constraint, numericPrecision);
-        return getGeometryTypeNameFromCode(geometryTypeCode);
-    }
-
-    public static String getGeometryTypeNameFromCode(int geometryTypeCode) {
-        return TYPE_MAP.get(geometryTypeCode);
+        return SFSUtilities.getGeometryTypeNameFromCode(geometryTypeCode);
     }
 }
